requirements.txt
PySide6==6.8.3
opencv-python>=4.9
numpy>=1.24
Pillow>=10.0

# Note: on macOS you may need to set QT_QPA_PLATFORM_PLUGIN_PATH to the PySide6 plugins/platforms dir
# Example:
# QT_QPA_PLATFORM_PLUGIN_PATH=$(python -c "import PySide6, os; print(os.path.join(os.path.dirname(PySide6.__file__), 'Qt', 'plugins', 'platforms'))") \
#   .venv/bin/python AnComicsViewer.py
